dragon-spell/challenge/requirements.txt
pillow
os
pathlib